Talk at workshop for Time Series Econometrics
This talk considers the estimation of aDFM using CVA type subspace methods. The main contribution is the suggestion of weighting matrices which lead to consistent estimators of the state space systems for stationary and integrated static factors process driven by q <= r dynamic factors (that is including the tall TF case).